+++ src/analyzer/protocol/rpc/MOUNT.cc
@@ -69,7 +69,7 @@ bool MOUNT_Interp::RPC_BuildCall(RPC_CallInfo* c, const u_char*& buf, int& n)
 	}
 
 bool MOUNT_Interp::RPC_BuildReply(RPC_CallInfo* c, BifEnum::rpc_status rpc_status,
-                                  const u_char*& buf, int& n, double start_time, double last_time,
+                                  const u_char*& buf, int& n, int64_t start_time, int64_t last_time,
                                   int reply_len)
 	{
 	EventHandlerPtr event = nullptr;
@@ -171,8 +171,8 @@ bool MOUNT_Interp::RPC_BuildReply(RPC_CallInfo* c, BifEnum::rpc_status rpc_statu
 	}
 
 Args MOUNT_Interp::event_common_vl(RPC_CallInfo* c, BifEnum::rpc_status rpc_status,
-                                   BifEnum::MOUNT3::status_t mount_status, double rep_start_time,
-                                   double rep_last_time, int reply_len, int extra_elements)
+                                   BifEnum::MOUNT3::status_t mount_status, int64_t rep_start_time,
+                                   int64_t rep_last_time, int reply_len, int extra_elements)
 	{
 	// Returns a new val_list that already has a conn_val, and mount3_info.
 	// These are the first parameters for each mount_* event ...
